requirements.txt
pyyaml==5.3
pandas==1.0.5
boto3==1.13.12
pymongo==3.10.1
xarray==0.15.1
scipy==1.3.2
tethys-utils==0.0.15
schedule==0.6.0
